// backend/src/parcels/interfaces/parcel.interface.ts
import { WeightCategory, ParcelStatus } from "generated/prisma";


export interface IParcel {
  id: string;
  senderId: string;
  receiverId: string;
  assignedCourierId?: string | null;

  receiverName: string;
  receiverPhone: string;
  pickupAddress: string;
  pickupLat: number;
  pickupLng: number;
  destination: string;
  destinationLat: number;
  destinationLng: number;
  weightCategory: WeightCategory;
  status: ParcelStatus;

  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date | null;
}